refactor(game): split 674-line recruit page into focused files under 250 lines (#734)
Reduces app/game/recruit/page.tsx from 674 lines to 51 by extracting:

- 8 presentation components into app/game/recruit/_components/
  - RecruitView — composed body with picker state + derived memos
    (military tiles, threat ranking, cap math, plan preview)
  - BulkRecruitPanel — recruit form (unit type / count / target)
  - StatGrid — 8-cell stats row
  - PlanPreview — auto-route preview
  - ProgressBar — recruit progress + artifact count
  - MilitaryTilesList — your military tiles
  - ReportLog — field reports
  - RecruitGates — Loading/SignIn/Enlist

- 5 lib modules into app/game/recruit/_lib/
  - threat-priority-plan — pure cycle distribution, with unit tests
  - use-recruit-data — player + tiles + borderTiles + owners
  - use-recruit-action — handleRecruit + progress + reports
  - types — response shapes, OwnerSummary, RecruitProgress
  - constants — UNITS_PER_CYCLE, TURNS_PER_CYCLE, RARITY_TEXT

The _components and _lib prefixes are Next.js private folders, so the new
structure does not introduce any new routes. page.tsx is now a thin gate
that picks one of {Loading, SignIn, Enlist, RecruitView} based on the
data-hook output.

Adds unit tests for buildThreatPriorityPlan covering empty inputs, total-
conservation, front-loading, single-tile routing, and remainder-allocation.

No behavior change. All 1589 tests pass (+8 new); tsc clean; lint clean.
